backend/requirements.txt
appdirs==1.4.4
arabic-reshaper==2.1.4
asgiref==3.3.4
async-generator==1.10
attrs==21.2.0
beautifulsoup4==4.11.1
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
colorama==0.4.4
colorclass==2.2.2
compressed-rtf==1.0.6
cryptography==36.0.0
cycler==0.10.0
defusedxml==0.7.1
distlib==0.3.1
Django==3.2.5
django-allauth==0.51.0
django-axes==5.13.1
django-bootstrap-form==3.4
django-cors-headers==3.7.0
django-ipware==3.0.2
django-markdown-deux==1.0.5
django-rest-framework==0.1.0
django-tinymce==3.3.0
djangorestframework==3.12.4
easygui==0.98.3
ebcdic==1.1.1
email-reply-parser==0.5.12
et-xmlfile==1.1.0
extract-msg==0.36.4
filelock==3.0.12
flake8==3.9.2
future==0.18.2
googlemaps==4.4.5
gunicorn==20.1.0
h11==0.12.0
html5lib==1.1
idna==2.10
IMAPClient==2.3.1
kiwisolver==1.3.1
lark-parser==0.12.0
lxml==4.6.3
markdown2==2.4.0
matplotlib==3.4.1
mccabe==0.6.1
msoffcrypto-tool==5.0.0
numpy==1.20.2
oauthlib==3.2.0
olefile==0.46
oletools==0.60.1
opencv-python==4.5.3.56
openpyxl==3.0.10
outcome==1.1.0
pandas==1.4.2
pcodedmp==1.2.6
Pillow==8.2.0
pipenv==2022.11.11
psycopg2==2.8.6
pycodestyle==2.7.0
pycparser==2.21
pyflakes==2.3.1
PyJWT==2.4.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
python-docx==0.8.10
python3-openid==3.2.0
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
pywin32==304
reportlab==3.5.67
requests==2.25.1
requests-oauthlib==1.3.1
RTFDE==0.0.2
selenium==4.1.0
simplejson==3.17.2
six==1.15.0
sniffio==1.2.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sortedcontainers==2.4.0
soupsieve==2.2.1
sqlparse==0.4.1
torch==1.9.1
torchvision==0.10.1
tqdm==4.64.0
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==3.10.0.2
tzdata==2022.5
tzlocal==4.2
uptime==3.0.1
urllib3==1.26.4
virtualenv==20.4.6
virtualenv-clone==0.5.7
webdriver-manager==3.6.1
webdrivermanager==0.10.0
webencodings==0.5.1
win-unicode-console==0.5
WMI==1.5.1
wsproto==1.0.0
xhtml2pdf==0.2.5